Working implementation of membership transfer based on creation of soft credit.
Unfortunately, this works only in the case of adding a soft contribution to an existing
set of membership/contribution records. In the case of a public form submission, which
is the use case we're really interested in, the membershipPayment record doesn't exist
yet (because the soft contribution is created first), so we aren't able to link the
soft contribution to the membership yet.